requirements.txt
yfinance>=0.2
requests>=2.31
openai>=1.0
twilio>=9.0
python-dotenv>=1.0
apscheduler>=3.10
